refactor App to clean up code by adding Form component to handle input data
